requirements.txt
Flask==3.1.3
Flask-CORS==6.0.2
google-generativeai==0.8.6
google-genai==1.66.0
gunicorn==23.0.0
